+++ tests/tabContainer.test.ts
@@ -8,6 +8,7 @@ import {
 } from "@std/testing/asserts";
 import Tab from "/tab.js";
 import { ensureAllTabsAvailability, TabContainer } from "/tabContainer.js";
+const currentDate = Date.now();
 
 function matchStorageToContainer(container: TabContainer) {
 	if (mockStorage.againWhySalesforce.length !== container.length) {
@@ -53,11 +54,21 @@ await Deno.test("TabContainer - Initialization", async (t) => {
 	await t.step("initializes with saved tabs from storage", async () => {
 		mockStorage.againWhySalesforce = [
 			{ label: "Test", url: "test-url" },
+			{
+				label: "Test2",
+				url: "test-url2",
+				[Tab.keyClickCount]: 3,
+				[Tab.keyClickDate]: currentDate,
+			},
 		];
 		const container = await TabContainer._reset();
-		assertEquals(container.length, 1);
+		assertEquals(container.length, 2);
 		assertEquals(container[0].label, "Test");
 		assertEquals(container[0].url, "test-url");
+		assertEquals(container[1].label, "Test2");
+		assertEquals(container[1].url, "test-url2");
+		assertEquals(container[1][Tab.keyClickCount], 3);
+		assertEquals(container[1][Tab.keyClickDate], currentDate);
 	});
 
 	await t.step(
@@ -86,10 +97,17 @@ await Deno.test("TabContainer - Tab Management", async (t) => {
 		assertEquals(lastTab.label, "New Tab");
 		assertEquals(lastTab.url, "new-url");
 		matchStorageToContainer(container);
-		assert(await container.addTab({ label: "New Tabb", url: "new-urll" }));
+		assert(
+			await container.addTab({
+				label: "New Tabb",
+				url: "new-urll",
+				[Tab.keyClickCount]: 9,
+			}),
+		);
 		const newLastTab = container.at(-1);
 		assertEquals(newLastTab.label, "New Tabb");
 		assertEquals(newLastTab.url, "new-urll");
+		assertEquals(newLastTab[Tab.keyClickCount], 9);
 		matchStorageToContainer(container);
 	});
 
@@ -98,16 +116,22 @@ await Deno.test("TabContainer - Tab Management", async (t) => {
 		assert(
 			await container.addTabs([
 				{ label: "Tab1", url: "url1" },
-				{ label: "Tab2", url: "url2" },
-				{ label: "Tab3", url: "url3" },
+				{ label: "Tab2", url: "url2", org: "test2" },
+				{ label: "Tab3", url: "url3", [Tab.keyClickCount]: 2 },
+				{ label: "Tab4", url: "url4", [Tab.keyClickDate]: currentDate },
 			]),
 		);
-		assertEquals(container.at(-3).label, "Tab1");
-		assertEquals(container.at(-2).label, "Tab2");
-		assertEquals(container.at(-1).label, "Tab3");
-		assertEquals(container.at(-3).url, "url1");
-		assertEquals(container.at(-2).url, "url2");
-		assertEquals(container.at(-1).url, "url3");
+		assertEquals(container.at(-4).label, "Tab1");
+		assertEquals(container.at(-3).label, "Tab2");
+		assertEquals(container.at(-2).label, "Tab3");
+		assertEquals(container.at(-1).label, "Tab4");
+		assertEquals(container.at(-4).url, "url1");
+		assertEquals(container.at(-3).url, "url2");
+		assertEquals(container.at(-2).url, "url3");
+		assertEquals(container.at(-1).url, "url4");
+		assertEquals(container.at(-3).org, "test2");
+		assertEquals(container.at(-2)[Tab.keyClickCount], 2);
+		assertEquals(container.at(-1)[Tab.keyClickDate], currentDate);
 		matchStorageToContainer(container);
 	});
 
@@ -119,6 +143,13 @@ await Deno.test("TabContainer - Tab Management", async (t) => {
 		);
 		assert(
 			await container.addTab({ label: "New Tab", url: "unique-url2" }),
+		);
+		assert(
+			await container.addTab({
+				label: "New Tab",
+				url: "unique-url2",
+				org: "test-org2",
+			}),
 		);
 		await assertRejects(
 			async () =>
@@ -926,6 +957,36 @@ await Deno.test("TabContainer - Import", async (t) => {
 			2,
 		);
 		assertEquals(container.length, 3);
+
+		assertEquals(
+			await container.importTabs(
+				`[{"label":"hello","url":"nice-url4","${Tab.keyClickCount}":4},{"label":"orglabel","url":"orgurl4","org":"orgorg","${Tab.keyClickDate}":${currentDate}}]`,
+				{
+					importMetadata: true,
+				},
+			),
+			2,
+		);
+		assertEquals(container.length, 5);
+		assertEquals(container.at(-2)[Tab.keyClickCount], 4);
+		assertEquals(container.at(-2)[Tab.keyClickDate], undefined);
+		assertEquals(container.at(-1)[Tab.keyClickCount], undefined);
+		assertEquals(container.at(-1)[Tab.keyClickDate], currentDate);
+		assertEquals(
+			await container.importTabs(
+				`[{"label":"hello","url":"nice-url5","${Tab.keyClickCount}":4,"${Tab.keyClickDate}":${currentDate},"org":"test-org5"}]`,
+				{
+					importMetadata: false,
+				},
+			),
+			1,
+		);
+		assertEquals(container.length, 6);
+		assertEquals(container.at(-1).label, "hello");
+		assertEquals(container.at(-1).url, "nice-url5");
+		assertEquals(container.at(-1).org, "test-org5");
+		assertEquals(container.at(-1)[Tab.keyClickCount], undefined);
+		assertEquals(container.at(-1)[Tab.keyClickDate], undefined);
 	});
 
 	await t.step("does not import tabs from wrong JSON string", async () => {
@@ -1400,6 +1461,14 @@ await Deno.test("TabContainer - Update Tab", async () => {
 	matchStorageToContainer(container);
 	assert(
 		await container.updateTab({ url: "ManageUsers/home" }, {
+			[Tab.keyClickCount]: 1,
+			[Tab.keyClickDate]: currentDate,
+		}),
+	);
+	assertEquals(container[2][Tab.keyClickCount], 1);
+	assertEquals(container[2][Tab.keyClickDate], currentDate);
+	assert(
+		await container.updateTab({ url: "ManageUsers/home" }, {
 			org: "",
 			url: "",
 			label: "",
@@ -1433,22 +1502,22 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 		assert(
 			await container.updateTab(container[0], {
 				org: "a",
-				"click-count": 0,
-				"click-date": 0,
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
 			}),
 		);
 		assert(
 			await container.updateTab(container[1], {
 				org: "c",
-				"click-count": 2,
-				"click-date": 2,
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
 			}),
 		);
 		assert(
 			await container.updateTab(container[2], {
 				org: "b",
-				"click-count": 1,
-				"click-date": 1,
+				[Tab.keyClickCount]: 1,
+				[Tab.keyClickDate]: 1,
 			}),
 		);
 		// move around because they were already sorted
@@ -1495,22 +1564,22 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 		assert(
 			await container.updateTab(container[0], {
 				org: "a",
-				"click-count": 0,
-				"click-date": 0,
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
 			}),
 		);
 		assert(
 			await container.updateTab(container[1], {
 				org: "c",
-				"click-count": 2,
-				"click-date": 2,
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
 			}),
 		);
 		assert(
 			await container.updateTab(container[2], {
 				org: "b",
-				"click-count": 1,
-				"click-date": 1,
+				[Tab.keyClickCount]: 1,
+				[Tab.keyClickDate]: 1,
 			}),
 		);
 		matchStorageToContainer(container);
@@ -1551,21 +1620,21 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 					label: "flows",
 					url: "flows",
 					org: "e",
-					"click-count": 0,
-					"click-date": 0,
+					[Tab.keyClickCount]: 0,
+					[Tab.keyClickDate]: 0,
 				},
 				{
 					label: "assignment",
 					url: "assignment",
 					org: "a",
-					"click-count": 2,
-					"click-date": 2,
+					[Tab.keyClickCount]: 2,
+					[Tab.keyClickDate]: 2,
 				},
 				{
 					label: "field",
 					url: "field",
-					"click-count": 1,
-					"click-date": 1,
+					[Tab.keyClickCount]: 1,
+					[Tab.keyClickDate]: 1,
 				},
 			]),
 		);
@@ -1602,22 +1671,22 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 		assert(
 			await container.updateTab(container[0], {
 				org: "a",
-				"click-count": 0,
-				"click-date": 0,
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
 			}),
 		);
 		assert(
 			await container.updateTab(container[1], {
 				org: "c",
-				"click-count": 2,
-				"click-date": 2,
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
 			}),
 		);
 		assert(
 			await container.updateTab(container[2], {
 				org: "b",
-				"click-count": 1,
-				"click-date": 1,
+				[Tab.keyClickCount]: 1,
+				[Tab.keyClickDate]: 1,
 			}),
 		);
 		// now should all be random
@@ -1650,15 +1719,15 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 		assert(
 			await container.updateTab(container[0], {
 				org: "a",
-				"click-count": 0,
-				"click-date": 0,
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
 			}),
 		);
 		assert(
 			await container.updateTab(container[1], {
 				org: "c",
-				"click-count": 2,
-				"click-date": 2,
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
 			}),
 		);
 		// now should all be random
@@ -1691,15 +1760,15 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 		assert(
 			await container.updateTab(container[0], {
 				org: "a",
-				"click-count": 0,
-				"click-date": 0,
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
 			}),
 		);
 		assert(
 			await container.updateTab(container[1], {
 				org: "c",
-				"click-count": 2,
-				"click-date": 2,
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
 			}),
 		);
 		// now should all be random
@@ -1718,5 +1787,179 @@ await Deno.test("TabContainer - Sort Tabs", async (t) => {
 		assertEquals(container[0].org, "c");
 		assertEquals(container[1].org, "a");
 		assertEquals(container[2].org, undefined);
+	});
+
+	// todo
+	await t.step("By click-count ascending", async () => {
+		await container.setDefaultTabs();
+		assertEquals(container.length, 3);
+		assertEquals(container[0].url, "/lightning");
+		assertEquals(container[1].url, "/lightning/app/standard__FlowsApp");
+		assertEquals(container[2].url, "ManageUsers/home");
+		// move around because they were already sorted
+		container.moveTab({ label: "flows" }, { moveBefore: true });
+		// add orgs
+		assert(
+			await container.updateTab(container[0], {
+				org: "a",
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
+			}),
+		);
+		assert(
+			await container.updateTab(container[1], {
+				org: "c",
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
+			}),
+		);
+		// now should all be random
+		matchStorageToContainer(container);
+		assertFalse(container.isSorted);
+		assertEquals(container.isSortedBy, null);
+		assertFalse(container.isSortedAsc);
+		assertFalse(container.isSortedDesc);
+		// sort by org ascending
+		assert(
+			await container.sort({ sortBy: Tab.keyClickCount, sortAsc: true }),
+		);
+		matchStorageToContainer(container);
+		assert(container.isSorted);
+		assertEquals(container.isSortedBy, Tab.keyClickCount);
+		assert(container.isSortedAsc);
+		assertFalse(container.isSortedDesc);
+		assertEquals(container[0][Tab.keyClickCount], undefined);
+		assertEquals(container[1][Tab.keyClickCount], 0);
+		assertEquals(container[2][Tab.keyClickCount], 2);
+	});
+
+	await t.step("By click-count descending", async () => {
+		await container.setDefaultTabs();
+		assertEquals(container.length, 3);
+		assertEquals(container[0].url, "/lightning");
+		assertEquals(container[1].url, "/lightning/app/standard__FlowsApp");
+		assertEquals(container[2].url, "ManageUsers/home");
+		// move around because they were already sorted
+		container.moveTab({ label: "flows" }, { moveBefore: true });
+		// add orgs
+		assert(
+			await container.updateTab(container[0], {
+				org: "a",
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: 0,
+			}),
+		);
+		assert(
+			await container.updateTab(container[1], {
+				org: "c",
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: 2,
+			}),
+		);
+		// now should all be random
+		matchStorageToContainer(container);
+		assertFalse(container.isSorted);
+		assertEquals(container.isSortedBy, null);
+		assertFalse(container.isSortedAsc);
+		assertFalse(container.isSortedDesc);
+		// sort by org descending
+		assert(
+			await container.sort({ sortBy: Tab.keyClickCount, sortAsc: false }),
+		);
+		matchStorageToContainer(container);
+		assert(container.isSorted);
+		assertEquals(container.isSortedBy, Tab.keyClickCount);
+		assertFalse(container.isSortedAsc);
+		assert(container.isSortedDesc);
+		assertEquals(container[0][Tab.keyClickCount], 2);
+		assertEquals(container[1][Tab.keyClickCount], 0);
+		assertEquals(container[2][Tab.keyClickCount], undefined);
+	});
+
+	const stepCurrentDate = Date.now() + 1000;
+	await t.step("By click-date ascending", async () => {
+		await container.setDefaultTabs();
+		assertEquals(container.length, 3);
+		assertEquals(container[0].url, "/lightning");
+		assertEquals(container[1].url, "/lightning/app/standard__FlowsApp");
+		assertEquals(container[2].url, "ManageUsers/home");
+		// move around because they were already sorted
+		container.moveTab({ label: "flows" }, { moveBefore: true });
+		// add orgs
+		assert(
+			await container.updateTab(container[0], {
+				org: "a",
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: currentDate,
+			}),
+		);
+		assert(
+			await container.updateTab(container[1], {
+				org: "c",
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: stepCurrentDate,
+			}),
+		);
+		// now should all be random
+		matchStorageToContainer(container);
+		assertFalse(container.isSorted);
+		assertEquals(container.isSortedBy, null);
+		assertFalse(container.isSortedAsc);
+		assertFalse(container.isSortedDesc);
+		// sort by org ascending
+		assert(
+			await container.sort({ sortBy: Tab.keyClickDate, sortAsc: true }),
+		);
+		matchStorageToContainer(container);
+		assert(container.isSorted);
+		assertEquals(container.isSortedBy, Tab.keyClickDate);
+		assert(container.isSortedAsc);
+		assertFalse(container.isSortedDesc);
+		assertEquals(container[0][Tab.keyClickDate], undefined);
+		assertEquals(container[1][Tab.keyClickDate], currentDate);
+		assertEquals(container[2][Tab.keyClickDate], stepCurrentDate);
+	});
+
+	await t.step("By click-date descending", async () => {
+		await container.setDefaultTabs();
+		assertEquals(container.length, 3);
+		assertEquals(container[0].url, "/lightning");
+		assertEquals(container[1].url, "/lightning/app/standard__FlowsApp");
+		assertEquals(container[2].url, "ManageUsers/home");
+		// move around because they were already sorted
+		container.moveTab({ label: "flows" }, { moveBefore: true });
+		// add orgs
+		assert(
+			await container.updateTab(container[0], {
+				org: "a",
+				[Tab.keyClickCount]: 0,
+				[Tab.keyClickDate]: currentDate,
+			}),
+		);
+		assert(
+			await container.updateTab(container[1], {
+				org: "c",
+				[Tab.keyClickCount]: 2,
+				[Tab.keyClickDate]: stepCurrentDate,
+			}),
+		);
+		// now should all be random
+		matchStorageToContainer(container);
+		assertFalse(container.isSorted);
+		assertEquals(container.isSortedBy, null);
+		assertFalse(container.isSortedAsc);
+		assertFalse(container.isSortedDesc);
+		// sort by org descending
+		assert(
+			await container.sort({ sortBy: Tab.keyClickDate, sortAsc: false }),
+		);
+		matchStorageToContainer(container);
+		assert(container.isSorted);
+		assertEquals(container.isSortedBy, Tab.keyClickDate);
+		assertFalse(container.isSortedAsc);
+		assert(container.isSortedDesc);
+		assertEquals(container[0][Tab.keyClickDate], stepCurrentDate);
+		assertEquals(container[1][Tab.keyClickDate], currentDate);
+		assertEquals(container[2][Tab.keyClickDate], undefined);
 	});
 });